fix import + rename targetQ to constantQ
