+++ open-reac/src/main/java/com/powsybl/openreac/parameters/output/ReactiveSlackOutput.java
@@ -10,7 +10,7 @@ import com.powsybl.ampl.converter.AmplConstants;
 import com.powsybl.ampl.converter.AmplException;
 import com.powsybl.ampl.converter.AmplSubset;
 import com.powsybl.commons.util.StringToIntMapper;
-import com.powsybl.openreac.parameters.IncompatibleModelError;
+import com.powsybl.openreac.exceptions.IncompatibleModelError;
 
 import java.io.IOException;
 import java.nio.charset.StandardCharsets;
